+++ libs/appproxy/testutil/http_server.go
@@ -1,4 +1,4 @@
-package testserver
+package testutil
 
 import (
 	"io"
