+++ libs/appproxy/testutil/websocket_server.go
@@ -1,4 +1,4 @@
-package testserver
+package testutil
 
 import (
 	"bytes"
